move model interfaces
