+++ src/app/overview-report/facility-report/facility-report-utility-usage-table/facility-report-utility-usage-table.component.ts
@@ -2,7 +2,8 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { UtilityMeterdbService } from 'src/app/indexedDB/utilityMeter-db.service';
 import { IdbFacility, IdbUtilityMeter } from 'src/app/models/idb';
-import { OverviewReportService, ReportOptions, ReportUtilityOptions, ReportUtilitySummary } from '../../overview-report.service';
+import { ReportUtilityOptions, ReportUtilitySummary } from 'src/app/models/overview-report';
+import { OverviewReportService } from '../../overview-report.service';
 
 @Component({
   selector: 'app-facility-report-utility-usage-table',
