+++ src/app/overview-report/facility-report/facility-report.component.ts
@@ -1,7 +1,8 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { IdbFacility } from 'src/app/models/idb';
-import { OverviewReportService, ReportOptions } from '../overview-report.service';
+import { ReportOptions } from 'src/app/models/overview-report';
+import { OverviewReportService } from '../overview-report.service';
 
 @Component({
   selector: 'app-facility-report',
