+++ src/app/overview-report/front-page/front-page.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { OverviewReportService, ReportOptions } from '../overview-report.service';
+import { ReportOptions } from 'src/app/models/overview-report';
+import { OverviewReportService } from '../overview-report.service';
 
 @Component({
   selector: 'app-front-page',
